BUG PresentViewModel - Constructor was called several times
The reason was, that in ToolBarPresenterLegacy some textboxes had their DataContextes set to PresentViewModel in XAML - Code. So for each textbox a own PresentViewModel was created. Due to this, every time a staticpropertychanged event was called, all PresentViewModel- Handler where called (4 Times). Now the datacontext is bound to Ancestor (MainWindow). This way, the PresentViewModel is only created one times